refactor(integrations): extract ado/ sub-package — move 11 Azure DevOps files
Move 5 source + 6 test files from internal/tools/integrations/ to
internal/tools/integrations/ado/. Update package declaration to 'package ado'.
Drop 'azure_devops_' prefix from filenames (package name provides namespace).

Parent integrations/ package and registration.go will be fixed in
subsequent commits (Task A3).
