+++ internal/tools/integrations/ado/pipeline.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2026 gosharplite@gmail.com
 // SPDX-License-Identifier: MIT
 
-package integrations
+package ado
 
 import (
 	"bytes"
